refacor : getBasicFetch 파일 이름 수정 - getFetch
