+++ src/api/getFetch.ts
@@ -1,9 +1,7 @@
 import { instance } from './baseUrl';
 
-export const getBasicFetch = async <T>(endpoint: string, params = {}): Promise<T> => {
+export const getFetch = async <T>(endpoint: string, params = {}): Promise<T> => {
   const res = await instance.get(endpoint, { params });
   const data = res.data;
-  //TODO
-  console.log(data);
   return data;
 };
