+++ src/components/Home/PresentRanking/Item/useRankingItem.ts
@@ -1,4 +1,4 @@
-import { getBasicFetch } from '@src/api/getBasicFetch';
+import { getFetch } from '@src/api/getFetch';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import type { Goods } from '@src/types/Goods';
 import { useSuspenseQuery } from '@tanstack/react-query';
@@ -13,7 +13,7 @@ export const useRankingItem = ({
   const { data, isError } = useSuspenseQuery<Goods>({
     queryKey: ['ranking', { targetType, rankType }],
     queryFn: () =>
-      getBasicFetch<Goods>(BASIC_ENDPOINT.ranking, {
+      getFetch<Goods>(BASIC_ENDPOINT.ranking, {
         targetType,
         rankType,
       }),
