+++ src/components/Home/PresentTheme/Item/usePresentThemeFetch.ts
@@ -1,4 +1,4 @@
-import { getBasicFetch } from '@src/api/getBasicFetch';
+import { getFetch } from '@src/api/getFetch';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import type { Themes } from '@src/components/Home/PresentTheme/Item/ThemeType';
 import { useQuery } from '@tanstack/react-query';
@@ -6,7 +6,7 @@ import { useQuery } from '@tanstack/react-query';
 export const usePresentThemeFetch = () => {
   const { data, isError, isLoading } = useQuery<Themes>({
     queryKey: ['themes'],
-    queryFn: () => getBasicFetch<Themes>(BASIC_ENDPOINT.theme),
+    queryFn: () => getFetch<Themes>(BASIC_ENDPOINT.theme),
   });
   return {
     data,
